minor refactor and renaming
