+++ rules-legacy-quarkus-example/src/main/java/org/kie/kogito/legacy/LoanApplication.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito.queries;
+package org.kie.kogito.legacy;
 
 public class LoanApplication {
